+++ tests/models.py
@@ -44,6 +44,7 @@ class User(Model):
 class Email(Model):
     email_id = fields.IntField(primary_key=True)
     email = fields.CharField(max_length=200, db_index=True)
+    company = fields.CharField(max_length=100, db_index=True, unique=True)
     is_primary = fields.BooleanField(default=False)
     address = fields.CharField(max_length=200)
     users: fields.ManyToManyRelation[User] = fields.ManyToManyField("models.User")
@@ -86,6 +87,7 @@ class Product(Model):
     class Meta:
         unique_together = (("name", "type"),)
         indexes = (("name", "type"),)
+        managed = True
 
 
 class Config(Model):
@@ -101,6 +103,21 @@ class Config(Model):
     )
 
     email: fields.OneToOneRelation["Email"]
+
+    class Meta:
+        managed = True
+
+
+class DontManageMe(Model):
+    name = fields.CharField(max_length=50)
+
+    class Meta:
+        managed = False
+
+
+class Ignore(Model):
+    class Meta:
+        managed = False
 
 
 class NewModel(Model):